Allow dependabot to update GitHub Actions (and delete accidental `.github/.github` copy)
The CI failing on the recent merge is simply a deprecated
`actions/cache@v2` that is no longer allowed to be used.  We can use
`dependabot` to help suggest updates to GitHub Actions steps as well, in
addition to `cargo` crate dependencies.

Also delete an accidental, older copy-paste of `.github/` into
`.github/.github/`.