CI: Deduplicate `lint.yml` workflow from `ci.yml` workflow
We already have a CI workflow that runs a bunch of steps to build
and test the project including `cargo clippy` (but with warnings
allowed...) on every `push` and `pull_request`, which is duplicate with
the `lint.yml` that also runs `clippy` (though through deprecated and
soon-to-be-broken `actions-rs/clippy-check`)

Consolidate all the steps into the `ci.yml` workflow by making `cargo
clippy` more strict (most notably running on all workspace crates,
with all features enabled, while disallowing any and all warnings) and
running `rustfmt` here now (and not just on `nightly`)).

In a followup we should also merge `ci` and `nightly`.  The latter
only seems to test MacOS and Windows in addition which we'll have
to do directly in PRs when we start submitting Direct3D12 support
(and since DLSS doesn't provide a Metal backend, do we still want to
test MacOS via MoltenVK?).  And it's behind in that it doesn't test
`--no-default-features` and misses `sudo apt-get install libvulkan-dev`
in some places (which would need to be scoped to `ubuntu-latest` if it
turns out to be necessary in the first place).