Fix broken links to "inexistant" `nvngx-rs` crate
This repository is named `nvngx-rs` but hosts the `nvngx` and
`nvngx-sys` crate (good choice to not include `-rs` in the
inherently-Rust crate name!

Fix the links in the badges, and remove `Cargo.toml`'s `documentation`
field with these broken links entirely because those already default
to the correct `docs.rs/{package.name}` link when not set.