Upgrade `bindgen` to `0.72` to limit the bindings to MSRV `1.70`
`bindgen` always emits bindings to be compatible with some latest stable
when it was published.  Users can opt-out of that behaviour and ensure
the generated bindings are compatible with an older version; in this
case `bindgen` emits bindings for Rust `1.82` with `unsafe extern "C"`,
failing our MSRV `1.70` check.

This means that `bindgen` doesn't currently look at the Rust toolchain
version it was invoked with, nor what is specified in the surrounding
crates' `rust-version` (which could all potentially be ambiguous).