CI: Checkout submodules (`DLSS`) in every job that needs to build the crate (#21)
* CI: Checkout submodules (`DLSS`) in every job that needs to build the crate

The submodule needs to be checked out whenever the crate is built,
so that `bindgen` and `cc-rs` in the `build.rs` script of `nvngx-sys`
can find their sources.

* Bump MSRV to `1.70` for `bindgen 0.71` compatibility

Since `bindgen` is always ran in `build.rs`, rather than checking in its
results whenever the DLSS module or `bindgen` configuration is changed,
this crate has an impact on our MSRV as tested in CI.

* nvngx: Fix local dependency on `nvngx-sys`

Both a `path` and `version` can coexist and are recommended when
depending on other in-tree crates, so that modifications and freely
be made (to `nvngx-sys`) and directly translate into the other crate
(`nvngx`).

When publishing `nvngx`, `cargo` ensures it replaces the `nvngx-sys`
crate dependency with the specified version.

* Upgrade `bindgen` to `0.72` to limit the bindings to MSRV `1.70`

`bindgen` always emits bindings to be compatible with some latest stable
when it was published.  Users can opt-out of that behaviour and ensure
the generated bindings are compatible with an older version; in this
case `bindgen` emits bindings for Rust `1.82` with `unsafe extern "C"`,
failing our MSRV `1.70` check.

This means that `bindgen` doesn't currently look at the Rust toolchain
version it was invoked with, nor what is specified in the surrounding
crates' `rust-version` (which could all potentially be ambiguous).