build: Compile helpers with `cc` crate instead of open-coded `clang`+`ar`
To prepare for building on Windows, or with different compilers and
setups in general, use the `cc` crate which is able to detect and
configure a variety of compilers based on the host system as well as
(cross-compilation!) compiler flags.

The `.compile()` call automatically sets up the `rustc-link-lib` flags
for the linker to find and link against the generated `ngx_helpers`
static object.