sys: Checkin generated bindings and make `bindgen` optional
Not all projects like or accept `bindgen` in their tree, primarily
because of requiring a full clang/LLVM install on the development
machine making things harder to just run and work out of the box.  At
the same time the bindings for this header are generic and do not need
to be regenerated unless the bindgen config in `build.rs` changes, or
the submodule is updated.

For these reasons, make `bindgen` optional behind a feature and test in
CI that the generated output doesn't change.