Include FrameGeneration API.
Includes the frame generation API in the bindings.

The FG API is a C++ API, for some reason, as it uses an
inline-initialisation of the struct members, even though those are still
defined in the C way (typedef struct). To allow it to compile and link
properly, as well as provide the bindings for the new APIs, the CC is
tuned for treating the input as C++.

Also, bumps the MSRV to 1.71.