Add a baseline DLSS example (#42)
* Make include directory lowercase

* Show error on wrong path

* Add example

* Add mini init example

* Remove log

* Add allocator

* Add roundtrip

* Add motion vector image and depth image

* Add DLSS upscale

* Clean code

* Add some form of auto sync

* Fix some barriers

* Remove unneeded names

* Cleanup code

* Format

* Fix numbering

* Remove random comments

* Remove borrow

* sys: Only capitalize Vulkan SDK `Include` directory when targeting Windows

While less likely because of standard include, binary and library
directories, the `VULKAN_SDK` environment variable can still be set
on Unix/Linux systems (i.e. when downloading the SDK tarball manually
and pointing the variable to it rather than installing via a package
manager) and would erroneously use the inexistent capitalized `Include`
directory for Windows, when compiling the ""helper"" file.  Which was
furthermore inconsistently **lowercase** in header generation.

By looking at `CARGO_CFG_TARGET_OS` we can consistently use a
capitalized directory when compiling *on Windows* (not that important
as it is usually configured to be case-insensitive) or *cross-compiling
from Linux for Windows*.  In the latter case however, it's unclear if
users are supposed to piece together a Windows-like Vulkan SDK with
capitalized `Include` folder given that there's no Windows SDK tarball,
or reuse the Vulkan SDK tarball (or package) for Linux given that the
header _contents_ are identical and usable across platforms based on
preprocessor directives?

* sync Cargo.lock with traverse/master and update Cargo.lock with needed dependencies for the example

---------

Co-authored-by: Marijn Suijten <marijn@traverseresearch.nl>