sys: Add architecture check and document the `cfg()` to use
NVIDIA only provides Linux and Windows libraries of DLSS, and only for
the `x86_64` architecture.  We already error out if the `target_os` is
something different (as this repository implements zero fallbacks and
will likely fail linking) but didn't yet check the `target_arch`.

Secondly, document in the `README` how downstream crates should depend
on `nvngx` with a `cfg()` such that it's not compiled on unsupported
targets.